feat: update forgot_password_two.dart
