--- lib/auth/onboarding/intro.dart
@@ -1,45 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:get/get.dart';
-import 'package:hearth/auth/onboarding/onboarding.dart';
-import 'package:hearth/utils/colors.dart';
-
-
-class IntroScreen extends StatefulWidget {
-  const IntroScreen({super.key});
-
-  @override
-  State<IntroScreen> createState() => _IntroScreenState();
-}
-
-
-class _IntroScreenState extends State<IntroScreen> {
-  @override
-  void initState() {
-    super.initState();
-    Future.delayed(const Duration(seconds: 3), () {
-      Navigator.pushReplacement(
-        context,
-        MaterialPageRoute(builder: (context) => OnboardingScreen()),
-      );
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Container(
-        color: primaryColor,
-        width: double.infinity,
-        child: const Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          crossAxisAlignment: CrossAxisAlignment.center,
-          children: [
-            Image(image: AssetImage("assets/images/logo.png")),
-            SizedBox(height: 20.8,),
-            Image(image: AssetImage("assets/images/hearth.png"))
-          ],
-        ),
-      ),
-    );
-  }
-}
