--- lib/auth/onboarding/onboarding.dart
@@ -1,15 +1,0 @@
-import 'package:flutter/material.dart';
-
-class OnboardingScreen extends StatefulWidget {
-  const OnboardingScreen({super.key});
-
-  @override
-  State<OnboardingScreen> createState() => _OnboardingaScreenState();
-}
-
-class _OnboardingaScreenState extends State<OnboardingScreen> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold();
-  }
-}
