+++ lib/ui/auth/forgot_password.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:get/get_core/src/get_main.dart';
-import 'package:hearth/ui/onboarding/forgot_pasword_two.dart';
+import 'package:hearth/reuseable_widgets/custom_back_button.dart';
+import 'package:hearth/ui/auth/forgot_password_two.dart';
 import 'package:hearth/utils/colors.dart';
 import 'package:hearth/utils/text_style.dart';
