+++ lib/ui/auth/forgot_password_two.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:get/get_core/src/get_main.dart';
-import 'package:hearth/ui/onboarding/reset_complete.dart';
+import 'package:hearth/reuseable_widgets/custom_back_button.dart';
+import 'package:hearth/ui/auth/reset_complete.dart';
 import 'package:hearth/utils/colors.dart';
 import 'package:hearth/utils/text_style.dart';
