+++ lib/ui/auth/verification.dart
@@ -1,6 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:get/get_core/src/get_main.dart';
+import 'package:hearth/reuseable_widgets/custom_back_button.dart';
+import 'package:hearth/reuseable_widgets/custom_popup.dart';
 import 'package:hearth/utils/colors.dart';
 import 'package:hearth/utils/text_style.dart';
 import 'package:pinput/pinput.dart';
